backend: add user model
This is my initial implementation. `type` field is a bit incomplete
at the moment. I couldn't be sure if I should give the number 4 to
moderators or add a new boolean field `isModerator`. Since ER table
shows only `type` (renamed from `flags`, because name flag is ambiguous),
it remained like that.

I am not sure if we need to define the default values right here, but
they also can be added later.

Some fields must be mandatory but to keep it simple for now, I didn't go
into much detail.