fix.backend.models: fix annotations, field names, relationships
I realized we can benefit from using bi-directional relationship
here. It would definetly makes easier to access from one entity to
another. It's not necessary for all the fields, but for some of
them it would be beneficial. For example, vote <-> review.

I set the fetching type to EAGER all the time in @ManyToOne
relationships, although after v0.1, we can address the perfomance
issues.