backend.models.user: connect with product
I tried to use ProductRepository to fetch all of the products that
belongs to passed owner, but I ended up with NonUniqueResultsExc.
I couldn't couldn't solve it, so I had to make this relationship
bidirectional. To be honest, good thing that error came up. One
way or another we were going to do it anyway.