backend: use constructors or model methods to set model fields
After validation of data has been done in the controller, it must
be injected to the relevant model. And if we are going to create
a new object of model, then why not use its constructors? Moreover
when a model needs to get updated, setFrom<>Dto() methods handles
this without bothering the logic going on at the controller.

At the moment, some of the constructors of User model are not
being used. I deliberately leaved them to offer different options.
They can be scratched off if the team decides to.

[fmt]: format code