refactor: use 404 page, validate results of db queries
Spring Jpa design is so awful that they return a null value if the
database can't find something. That's why we need to check whether
the response is null or not everytime we use the repository. I
looked for if we can somehow customize the repository/jpa so that
they can throw an exception instead of returning null, but no
viable result. Therefore I added missing control blocks. While
doing that, I also made small adjustments on the authorization and
authentication logic so that there are fewer database query.

Sometimes ORMs are smart, they don't immediately hit the database.
Maybe that was the case here, I didn't make a research, but still
there were situtations where we queried for displayedUser twice.