backend: init ReplyReviewDto
We have decided not to use seperate DTOs for a model, except User.
However putting merchantReplyBody into Review model kinda required
me to use different dto for this. It's still possible to use
existing ReviewDto for merchant replies with small tweaks, but it
doesn't feel right to hide many other inputs which aren't relevant
to the replies. One other choice would be obviously creating a new
model for replies, but pffft :eyes