backend: set maximum number of images for a product
Restricting the displayOrder at ProductImage is silly (blame me).

I thought of doing a similar thing for user profile images as well
but using list of multipartfile for something we really don't
accept and controlling them if its size is greater than one didn't
click for me. If users choose more than one file by hacking the
html, it doesn't matter, only one of them is saved. In firefox,
specifically the first one that is selected is being saved.

I realized that Spring boot has limits for the size of the files,
requests. I was trying to upload 17 images to test the maximum
limit, program failed to render a view. I really don't want to
tinker the spring settings right now, it shouldn't be priority.
But hey I found a bug!