backend: give permission to admins for deleting products and reviews
Why can't admins edit reviews and products? Well, that would be
really awful in real life. Someone can write something from your
account. Admins would have the chance to abuse this. That's the
same case with creating reviews and products. Ideally, when an
admin deletes it, system should send an email. I introduced
EmailService in #114, hopefully it's going to be usable when it
gets merged. However, it shouldn't be a priority. And it's not
even in the requirements so.. yeah. It would have been nice to
include such things in system requirements documentation.

I also did another change in this commit while I was touching
the authentication and authorizations. When user exist in session,
but not in database was considered impossible. Actually it's not.
User can be deleted by admins while she is in session, although
we don't have this feature yet. Proper way of handling this is
doing logout and redirecting to login page. Plus, an informative
message can be given here, that would be nice.