fix.backend: skip empty files
SpringBoot weirdly adds an empty file to the list of MultipartFile
if users doesn't upload any. Therefore I added a control block to
skip when the file is empty.

ContentType is application/octet-stream for the empty file.

Fixes #119

Here, I realized that we don't do any check whether the uploaded
files are images. We can do it by checking the ContentType but I'm
not sure if it would be a certain solution. It doesn't break the
program, if they upload such file then they can't see the content.
Therefore I don't see this as priority since there are 5 days left
for final presentation.